requirements.txt
fastapi
uvicorn
supabase
tiktoken
openai
neo4j
pydantic
requests
python-dotenv
numpy